--- src/main/java/com/algo/common/DisableSecurityConfig.java
@@ -1,25 +1,0 @@
-package com.algo.common;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Slf4j
-@Configuration
-@EnableWebSecurity
-@ConditionalOnProperty(name = "algomi.security.enable", havingValue = "auto")
-public class DisableSecurityConfig {
-
-  @Bean
-  public SecurityFilterChain securityFilterChain(HttpSecurity http)
-      throws Exception {
-    http.formLogin(AbstractHttpConfigurer::disable);
-    http.csrf(AbstractHttpConfigurer::disable);
-    return http.build();
-  }
-}
