+++ src/main/java/com/algo/exception/ExceptionControllerAdvice.java
@@ -1,4 +1,4 @@
-package com.algo.others;
+package com.algo.exception;
 
 import org.springframework.web.bind.annotation.ControllerAdvice;
